+++ CHANGELOG.md
@@ -4,6 +4,16 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 
 
+
+## [1.11.2 / 5.66.2] - 2023-08-
+
+### Added
+- added update warning when updating with an expired personal certificate
+
+### Fixed
+- fixed subscription certificate recognition issue
+
+
 ## [1.11.1 / 5.66.1] - 2023-08-31
 
 ### Added